Replace v1 annotation references with v2
Move AllAllowedAnnotations to v2 package and deprecate the v1 reference.
Use v2 annotations directly instead of going through the deprecated v1 package.
Replace goresctrl rdt import with plain string constant to keep v2 as a
constants-only package.
